Controller Input (WIP4)
I have made the audio menu work with a controller correctly as it used to scroll to fast to chose and only went up to a certain point. The only issue is that you can't reopen the audio menu once closing unless you use a controller and click the options button.