Controller Input (WIP8)
Shit forgot to commit last hour. So for the last thing I added the controller stuff to the character customization. Currently for what was meant to be the 9th commit I am working on controls for the map. It kind of works but there is an issue of going left and right I dislike.